main.py
from PyQt5.QtWidgets import QApplication, QFileDialog, QCheckBox, QVBoxLayout, QWidget, QListView, QAbstractItemView, QMessageBox
from PyQt5.QtCore import QStringListModel
from google_auth_oauthlib.flow import InstalledAppFlow
from googleapiclient.discovery import build
from googleapiclient.http import MediaFileUpload, MediaIoBaseDownload
from googletrans import Translator
from docx import Document
import io
import os

# Create a translator object
translator = Translator()

# Create a Qt application
app = QApplication([])

# Set up the Google Drive API
SCOPES = ['https://www.googleapis.com/auth/drive']
flow = InstalledAppFlow.from_client_secrets_file('client_secrets.json', SCOPES)
creds = flow.run_local_server(port=0)
service = build('drive', 'v3', credentials=creds)

def list_files_in_folder(folder_id):
    results = service.files().list(
        pageSize=10, q=f"'{folder_id}' in parents").execute()
    items = results.get('files', [])
    return items

def download_file(file_id, filepath):
    request = service.files().get_media(fileId=file_id)
    fh = io.FileIO(filepath, 'wb')
    downloader = MediaIoBaseDownload(fh, request)
    done = False
    while done is False:
        status, done = downloader.next_chunk()

def translate_file(filepath):
    # This function would use the googletrans library to translate the file
    pass

def upload_file(filename, filepath, mimetype):
    file_metadata = {'name': filename}
    media = MediaFileUpload(filepath, mimetype=mimetype)
    file = service.files().create(body=file_metadata, media_body=media, fields='id').execute()

def main():
    # List files in the user's Google Drive
    results = service.files().list(pageSize=1000).execute()
    items = results.get('files', [])
    file_names = [item['name'] for item in items]

    # Create a list view with checkboxes for the user to select files
    model = QStringListModel(file_names)
    view = QListView()
    view.setModel(model)
    view.setSelectionMode(QAbstractItemView.MultiSelection)
    view.show()

    # Wait for the user to close the list view
    app.exec_()

    # Get the selected files
    selected_indexes = view.selectedIndexes()
    selected_files = [model.data(index) for index in selected_indexes]

    # Download, translate, and upload each selected file
    for file_name in selected_files:
        file_id = next(item['id'] for item in items if item['name'] == file_name)
        download_file(file_id, file_name)
        translate_file(file_name)
        upload_file(file_name, file_name, 'application/vnd.openxmlformats-officedocument.wordprocessingml.document')

    # Show a message box when the translation is done
    msg = QMessageBox()
    msg.setIcon(QMessageBox.Information)
    msg.setText("Translation completed!")
    msg.exec_()

if __name__ == "__main__":
    main()
    app.exec_()
