requirements.txt
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
googletrans==4.0.0-rc1
python-docx
PyQt5